Refactor Aca class and add MaAca tax logic
