+++ programs/programs/ma/pe/tax.py
@@ -1,4 +1,4 @@
-from programs.programs.federal.pe.tax import Eitc
+from programs.programs.federal.pe.tax import Aca, Eitc
 import programs.programs.policyengine.calculators.dependencies as dependency
 from programs.programs.policyengine.calculators.base import PolicyEngineTaxUnitCalulator
 
@@ -19,3 +19,11 @@ class MaChildFamilyCredit(PolicyEngineTaxUnitCalulator):
         *dependency.irs_gross_income,
     ]
     pe_outputs = [dependency.tax.MaChildFamilyCredit]
+
+
+class MaAca(Aca):
+    pe_inputs = [
+        *Aca.pe_inputs,
+        dependency.household.MaStateCodeDependency,
+        dependency.household.ZipCodeDependency,
+    ]
